requirements.txt
# OmniPublish v2.0 dependencies
Pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.0
pycryptodome>=3.19.0
requests>=2.31.0
